Moved the odds/luck stuff into a new command called !odds. Fixed sabertooth not exporting hurt triggers. Added titles to the images.
